requirements.txt
# Python dependencies for NYTimesBestsellerReporter
requests>=2.31.0
python-dotenv>=1.0.0
